[runtime] Fix bugs exposed in Float16Array tests (#153)
## Summary

Fix two bugs exposed by test262 due to the addition of Float16Array
tests.

1. We should use wrapping arithmetic during `f64_to_f16` to avoid
failures in debug mode.
2. When transferring an ArrayBuffer we were not handling the case where
the old and new ArrayBuffers have different length. We were always using
the old ArrayBuffer as the backing store in this case without resizing
and potentially zeroing the new range of the array. Instead if the
length is changing on transfer we should always copy to a new array,
zeroing the new range if the ArrayBuffer is growing.

## Tests

All test262 tests now pass in both release mode and GC stress test mode.