[common] Use forked indexmap crate that adds allocator API support (#161)
## Summary

We plan on allocating the AST using a bump pointer allocator. This means
that all collections in the AST must use the allocator API. We can use
the `allocator_api2` crate for simple built-in collections like `Vec`
but all more complex, non-std, and custom collections must find a way to
support the allocator API.

The hardest of these collections is the third party `indexmap` crate
which does not support the allocator API, and [does not plan
to](https://github.com/indexmap-rs/indexmap/issues/135#issuecomment-1063115749)
until the allocator API is stabilized. We currently use `IndexMap` in
AST scope nodes.

Unfortunately there does not appear to be an `IndexMap` alternative
supporting the allocator API. Instead we took the step of forking
`indexmap` and patching the latest version (2.7.1) with support for the
allocator API across almost the entire API surface at
https://github.com/Hans-Halverson/indexmap-allocator-api. [This is the
patch](https://github.com/Hans-Halverson/indexmap-allocator-api/commit/4e33c59c43486136394f04708fb0de7fc3825fe5)
that was used.

The forked `indexmap-allocator-api` is then used as a drop-in
replacement for the `indexmap` crate (implicitly using the global
allocator), and can be created with explicit allocators in future PRs.

## Tests

All tests pass.