[parser] Prepare heap allocated values in AST to support allocator API (#162)
## Summary

We plan on allocating the AST using a bump pointer allocator. This means
that all collections in the AST must use the allocator API. We can use
the `allocator_api2` crate for simple built-in collections like `Vec`
but all more complex, non-std, and custom collections must find a way to
support the allocator API.

This PR performs some preparation, identifying heap allocated values in
the AST and updating them to support passing a custom allocator.

- `BigInt` from the `num_bigint` crate uses a `Vec` to store digits but
has no allocator API support. Instead when storing in AST we decompose
into a sign and `Vec`, then reconstruct the `BigInt` when necessary.
- `HashMap` and `HashSet` from the standard library do not support the
allocator api. Luckily the standard `hashbrown` crate does so we can
migrate to it for all cases in the AST.

## Tests

All tests pass.