[parser] AST uses Wtf8String instead of String in all nodes (#163)
## Summary

We plan on allocating the AST using a bump pointer allocator. This means
that all heap allocations in the AST must use the allocator API. We can
use the `allocator_api2` crate for simple built-in collections like
`Vec` but all more complex, non-std, and custom collections must find a
way to support the allocator API.

The built-in `String` type does not support custom allocators. Luckily
our custom `Wtf8String` can easily be modified to use a custom
allocator, and most strings in the AST are already `Wtf8String`s. This
PR refactors all remaining instances of `String` to `Wtf8String` in the
AST, RegExp AST, and AST scope nodes.

This is for the most part a straightforward migration. A `Wtf8String`
can be created by consuming a `String` and serves as a strictly weaker
representation - i.e. may contain unpaired surrogate code points. This
slight weakening is not meaningful anywhere we were using the built-in
`String` type.

Note that we now represent global built-in string names as `static
LazyLock` instead of `&'static str`.

## Tests

All tests pass.