[common] Use jemalloc as global allocator (#165)
## Summary

Looking at flamegraphs of performance a large amount of time is spent in
system malloc/free/madvise calls. Let's switch to using jemalloc on
supported platforms and it performs much better in practice.

For example when running the entire integration test suite with `cargo
brimstone -r`, using jemalloc reduced the runtime from ~6.6s to ~5.4s.

This also prevents a regression when switching to an arena allocator for
the AST, which appeared to be hitting slow madvise calls and
significantly regressed performance.

## Tests

All tests pass.