[parser] Use arena allocator for AST (#166)
## Summary

Building the AST requires a large number of small allocations that all
have the same lifetime. This is a natural choice for arena allocation,
allowing us to perform fewer allocations with the global allocator. We
use the standard `Bumpalo` crate for our arena allocator along with the
stable allocator API in `allocator_api2`.

This PR switches all AST types to use data structures parameterized by
our new arena allocator. This required a lot of refactoring and
threading lifetimes through the parser and analyzer.

- `ParseContext` owns the AST and must be created from the soruce before
parsing
- Analysis now consumes the `ParseProgramResult` and returns an
`AnalyzedProgramResult` which is used for bytecode generation
- Parameterized `Wtf8String` with an allocator
- Created a `Wtf8Str` slice analogous to `&str`

Results:

- Running the entire integration test suite is ~2% faster (`cargo
brimstone-test -r`)
- Larger parser benchmarks (`acorn`, `react`, `typescript`) are 2-18%
slower but we are optimistic this can be eliminated once future work is
complete.

Future work:

- We guarantee that the entire AST (including the scope tree and all
referenced fields) are allocated with the same arena allocator. This
will allow us to avoid dropping the AST and instead forget the AST and
drop the underlying arena (not yet implemented).
- Minimize copies to `Wtf8String` in the AST, instead use `Wtf8Str` to
directly reference slices of the source string
- Use direct references and slices in the AST instead of `AstBox` and
`AstVec` to avoid storing duplicate copies of allocator

## Tests

All tests pass.