[parser] Use custom AstBox and AstSlice type (#167)
## Summary

Shrink the size of the AST by using a custom `AstBox` and `AstSlice`
type which do not contain a reference to the allocator. The switch to
arena allocation for the AST in
https://github.com/Hans-Halverson/brimstone/pull/166 introduced some
size regressions due to the use of `allocator_api2::{Box, Vec}` which
contain a reference to the allocator.

In addition to removing the unnecessary reference to the allocator,
`AstSlice` cannot change in size and does not hold a capacity, halving
it's size.

This speeds up parsing both due to the reduced memory use and due to no
longer running destructors for the AST. The changes for some parsing
benchmarks:
- `fixtures/acorn.js > parse`: -5.4%
- `fixtures/react.js > parse`: -4.6%
- `fixtures/typescript.js > parse`: -5.9%

## Tests

All tests pass.