[parser] AST nodes use AstStrs which directly reference source without copying when possible (#169)
## Summary

The AST currently holds many `AstString`s, which requires copying from
the source buffer into an owned `AstString`. Instead the AST should hold
`AstStr`s which are smaller and can directly reference the source buffer
without copying.

It is not always possible to directly reference the source buffer, for
example when escape sequences are encountered. In this case we must fall
back to a slow path where we build an owned `AstString` value by value.
Alternatively we start out in the fast path which will not allocate
until a problematic byte is encountered.

This causes significant parse time reductions across all benchmarks.
- `acorn`: -28.9%
- `react`: -31.5%
- `typescript`: -28.2%

## Tests

All tests pass.