[parser] Refactor Expression, Statement, and Pattern nodes to be enum of pointers to AST nodes (#173)
## Summary

The current AST structure results in a lot of wasted space. Expression,
Statement, and Pattern nodes are all represented as enums which take the
size of their largest variant, wasting space for the more commonly used,
smaller variants. In the past we kept the size down by placing the
largest variants behind a pointer, but let's go ahead and move to
placing all variants behind a pointer instead. This means that
Expression, Statement, and Pattern effectively become a typed pair of
the tag and a pointer to the node itself. Each variant now only needs to
allocate the precise size of the corresponding AST node, resulting in no
wasted space inside these enums.

This refactoring also reduces the amount of copying within the parser,
particularly in converting Expression to OuterExpression.

The reduction in memory use and copying results in ~5% parse time
reductions across all benchmarks.

## Tests

All tests pass.