[parser] Reduce size of Token from 56 to 24 bytes by arena allocating RegExp and Template tokens (#174)
## Summary

Tokens are currently much larger than they need to be: 56 bytes. This is
due to the inlining of RegExp and template literals in the token. Moving
these to be arena-allocated shrinks the size of tokens to 24 bytes.

We also must change the format of BigInt literal tokens. Instead of
creating the BigInt during lexing we instead return the slice of digits
and base of the BigInt. Then during parsing we parse the digits into a
BigInt.

The reduced token size decreases parse time in all parser benchmarks.

- `acorn`: -6.5%
- `react`: -7.6%
- `typescript`: -6.0%

## Tests

All tests pass.