[common] Jemalloc gated by default jemalloc feature (#176)
## Summary

Gate the use of jemalloc behind a `jemalloc` feature. This feature is
turned on by default, and can be used to disable the custom allocator.
This is needed in some cases e.g. when another crate sets the global
allocator, such as in
https://github.com/Hans-Halverson/brimstone-bench-javascript-parser-written-in-rust.

Also define a `no_jemalloc` feature which supersedes the `jemalloc`
flag. This allows easily disabling the use of jemalloc even when default
features are present.

## Tests

All tests pass, verify they continue to use jemalloc by default.