[runtime] Register all rust vtables with the runtime (#179)
## Summary

In preparation for serializing the heap we need some additional
structure around Rust vtables placed in the heap. This occurs due to a
couple sites using Rust trait objects for dynamic dispatch: object
method dispatch, modules, and typed arrays.

This PR registers all Rust vtables that appear in the heap to a top
level collection and pairs each with a value in the new `RustVtable`
enum. This allows you to map between the `RustVtable` enum and the true
value of the vtable pointer, meaning in the future we can serialize to
`RustVtable` and deserialize to the vtable pointer in the new address
space.

## Tests

All tests pass.