[runtime] Serialize heap and embed in binary, use to initialize heap at runtime (#181)
## Summary

Starting up the VM requires creating creating a heap and then
initializing a large number of objects to create all builtins. We can
greatly speed this up by initializing from a serialized version of the
heap instead of actually running all this builtin setup logic. This
serialized heap is generated at compile time and embedded in the binary
itself. The serialied heap to use is set in the `Options`.

The serialized form of the heap has all absolute pointers rewritten as
offsets from the start of the heap. In addition all root pointers in the
`Context` must also have been rewritten as offsets. To deserialize we
simply traverse the heap and root pointers and rewrite back to an
absolute pointer by adding the base pointer of the heap. Roots offsets
are saved in a vector following traversal order which is the same during
serialization and deserialization. We also now visit weak pointers
during generic `HeapVisitor` traversal so that they can be rewritten
during serialization/deserialization.

There are also some non-heap pointers contained within the heap, namely
Rust vtable pointers for dynamic dispatch. These are serialized to their
`RustVtable` enum value and then deserialized to the true value of the
vtable pointer in the loaded program. We also now must visit these Rust
vtable pointers during generic `HeapVisitor` traversal.

Embedding the serialized heap requires some changes to the build and
module system. The core library (aka the `common`, `parser`, and
`runtime` directories) have been converted into a `brimstone_core`
crate, which is consumed by the `brimstone` executable crate. There is a
new `brimstone_serialized_heap` crate which provides the default
serialized heap, which can be set as the default with
`brimstone_serialized_heap::init()`. The `brimstone_serialized_heap`
crate contains a `build.rs` script to create a Context using
`brimstone_core` as a build dependency, serializes it, then writes the
serialized heap into a codegenned file which is used by the crate.

This significantly improves startup time.
- A benchmark comparing manual `Context` initialization vs initializing
a `Context` from the embedded serialized form found a 90.6 vs 35.4 ms
runtime, or a 61% reduction
- Running all test262 and integration tests with `brimstone-test -r`
went from to 5.24 to 4.31 seconds, an 18% reduction


## Test Plan

All tests pass. Added a benchmark test for initializing the Context from
a serialized heap.