[bytecode] Minimize Wtf8String clones during bytecode generation
Minimize the number of Wtf8String clones performed during bytecode generation. The vast majority of these were unnecessary since direct Wtf8Str slices of the source can be used. In all cases either change to Wtf8Str if possible, otherwise change to Wtf8Cow so that owned and borrowed strings can be mixed.

There is one remaining string clone, but this only occurs when falling back to the default name for a function.

Full list of changes:

- Remove AnyStr now that only AnyStr::Wtf8 is used
- Switch to Wtf8Str in PendingFunctionNode::{Expression, Arrow, Constructor}
- Switch to Wtf8Cow in PendingFunctionNode::Method
- ByteFunctionGenerator uses Wtf8Cow for name

All tests pass.