[runtime] Heap serialization accounts for differences in heap layout due to features enabled (#187)
## Summary

There is a bug in our heap serialization system due to the fact that the
set of Cargo features enabled can change the heap layout, e.g. due to
differences in `HeapInfo` size when the `handle_stats` flag is enabled.
This breaks heap deserialization as the encoded offsets are from the
start of the heap (and the `HeapInfo` is stored at the start of the
heap), so changes to `HeapInfo` size should be added as an extra offset
while deserializing pointers.

We can fix this by storing the size of the `HeapInfo` in the serialized
heap, then comparing this to the size of the `HeapInfo` when
deserializing. The heap's base pointer is then adjusted by this
difference when deserializing.

## Tests

All tests pass. Reduced heap size in integration tests with
`handle_stats` feature now passes all tests instead of segfaulting.