[bytecode] Store intermediate object or array in temporary while building to avoid clobbering observable register (#196)
## Summary

When building object and array literals we start by creating an empty
object/array then evaluate and add its properties/elements. However
evaluating these properties/elements can throw, in which case we need to
make sure that the destination register for the object/array has not
been clobbered by the intermediate object/array value.

This is the same general pattern used in other cases that can lead to
observable register clobbering like logical assignments and template
expressions.

Note that we can make a small optimization to avoid the unnecessary
temporary if there are no properties or elements of the literal.

## Tests

- Added integration tests that catch this case.
- Added bytecode snapshot tests to confirm that temporary is only used
when needed.