[bytecode] Fix scope exit at end of block statement with abrupt completion (#198)
## Summary

During bytecode generation we only need to emit instructions to pop VM
scopes at the end of a block statement if the block statement has a
potential normal completion. If the block statement has an abrupt
completion then the pop instructions are handled separately (none for
return, restoring saved scope for throw, handled by jump target
machinery for break and continue).

However despite not needing to emit explicit pop instructions at the end
of the block we still need to unwind the scope within the bytecode
generator. Otherwise the simulated state of the VM scope chain in the
bytecode generator will not match the true VM scope chain at runtime
past this point.

## Tests

- Added an integration regression test for this case
- Added bytecode snapshot tests covering the VM scopes for block
statements with abrupt completions