[runtime] Don't call well-known Symbol methods for RegExp on primitive values (#201)
## Summary

Implement the normative change in
https://github.com/tc39/ecma262/pull/3009 where well-known Symbol
methods for RegExps are no longer called on primitive values in
`String.prototype.{match, matchAll, replace, replaceAll, search,
split}`. These methods now verify that the target value is a object
before looking for methods instead of just checking that the target
value is non-nullish.

## Tests

test262 tests for this case now pass.