[runtime] Rename for consistent terminology of heap items (#209)
## Summary

Heap items are intermittently referred to as "heap items" and "heap
objects". Let's prefer "heap items" to avoid confusion with JS objects
which only a subset of heap items. This required renaming a few other
objects (notably object descriptors and the object kind enum).

The full set of changes:

- Renamed "HeapItem" struct to "AnyHeapItem". This struct was meant to
represent the most generic heap item (containing only a descriptor) so
the new name is possibly more aligned with its use. In addition we want
to free up the name for:
- Renamed "HeapObject" trait to "HeapItem"
- Replaced all forms of "heap object" with "heap item"
- Replaced all forms of "object descriptor" with "heap item descriptor"
- Replaced all forms of "object kind" with "heap item kind"


## Tests

All tests pass.