[bytecode] Fix miscompilation for class methods with computed names sometimes generating discontiguous NewClass arguments
We are currently miscompiling class methods with computed names, when the computed name's expression is a direct use a binding defined earlier before the body of the class.

The NewClass instruction expects its arguments (containing methods and computed method names) to be in a contiguous range of registers. A discontiguous range was being generated in this case because we weren't enforcing that each computed name must be generated into a new temporary register. The binding was resolved to a direct reference of a local/argument register, breaking the contiguous range.

We now make sure to place computed name results in a temporary register. We also add a debug assertion that the NewClass arguments list is contiguous.

- Added regress test for this case, testing computed method names coming from a local and from an argument
- Added bytecode snapshot test