[common] Enable Annex B extensions by default (#222)
## Summary

We keep encountering Annex B syntax. In the spirit of compatibility
let's switch to supporting Annex B by default instead of requiring it to
be opt in with a cli flag.

A new Rust feature is used to gate whether Annex B is enabled by
default. You can still always manually configure Annex B regardless of
whether this feature is enabled. To my knowledge Annex B support is not
yet expensive, but if that changes we will consider conditional
compilation.

Note that we only have partial support for Annex B at the moment.

## Tests

Verified that `cargo run` no parses Annex B syntax but`cargo run
--no-default-features` does not parse Annex B syntax. All tests pass.