[test] Fuzz testing with fuzzilli (#226)
## Summary

Integrate with
[googleprojectzero/fuzzilli](https://github.com/googleprojectzero/fuzzilli)
using a fork with support added for brimstone:
https://github.com/Hans-Halverson/fuzzilli-brimstone.

Sets up a new crate for fuzz testing that communicates with the fuzzilli
testing harness as described in
https://github.com/googleprojectzero/fuzzilli/tree/main/Targets.

Fuzzilli needs to integrate with LLVM's
[SanitizerCoverage](https://clang.llvm.org/docs/SanitizerCoverage.html)
in order to perform profile guided fuzzing of the engine. Rust does not
directly offer support for integration with `SanitizerCoverage` but we
can do it ourselves with:
1. Compile [sanitizer
runtime](https://github.com/googleprojectzero/fuzzilli/blob/main/Targets/coverage.c)
ahead of time to a shared library and link to it.
2. Send the right incantation of RUSTFLAGS to manually enable the
`SanitizerCoverage` pass in LLVM. This will use symbols that link to the
former shared library.

Included script `./test/fuzz/build.sh` to build brimstone for fuzz
testing and install dependencies, then `./test/fuzz/run.sh` to run
fuzzilli. For now must be run locally but the next step is to set up
continuous fuzzing.

## Tests

Verify that `./test/fuzz/build.sh` and `./test/fuzz/run.sh` work and
start fuzzing the engine. See that the run script detects crashes after
running for a while!