[bytecode] Fix switch miscompile when breaking out of a switch with materialized scopes (#227)
## Summary

Fixes a miscompile caught by the fuzzer. We were previously emitting
instructions to pop all runtime scopes at the "join block" after the
switch statement. However the join block is actually the target of
breaks which which themselves also emit all necessary scope pops. This
means that when a break was hit with actual runtime scopes we would pop
too many times.

Instead we should not be popping scopes after the join block and instead
introduce a "switch body end" block which is where the scopes are popped
for normal switch statement completion. This matches how all other
breakable statements are handled, switch statements were the odd one
out.

## Tests

- Bytecode test for this case
- Regression test for case which previously crashed at runtime