[runtime] Function.prototype.bind preserves null prototypes (#230)
## Summary

When a function with a null prototype was bound by
`Function.prototype.bind` we were incorrectly defaulting the resulting
bound function's prototype to `Function.prototype` instead of preserving
the null prototype.

We fix this by changing
`BuiltinFunction::create_builtin_function_without_properties` to take
the value of the `[[Prototype]]` slot directly instead of defaulting the
`None` value to `Function.prototype`. This means we now must always
specify the prototype the two callsites:
- `BuiltinFunction::create` always uses `Function.prototype`
- `BuiltinFunction::intrinsic_constructor` always explicitly specifies
the prototype as an `Intrinsic` (without any defaults)

## Tests

- Added integration test confirming that null prototypes are preserved
- Fixes `kangax-es6/misc.bound-function-prototype/*` tests