[runtime] Refactor runtime to handle allocation failures safely with Result type
Currently an allocation error on the managed heap (aka an OOM) will immediately terminate the process.

Instead lets switch to making all allocations fallible and propagating a result type. This required threading through these new result types though all methods that could potentially allocate. This affected virtually every file in the runtime.

We keep types as specific as possible, meaning we maintain AllocResult types for as long as possible before converting to the more general EvalResult. EvalResults are now either a thrown JS value or an allocation error. Allocation errors are immediately propagated upwards in most cases.

This regresses performance slightly (seeing around 2-4% on octane benchmarks, but inconsistent). Hoping this is worth it for the safety of being able to gracefully handle allocation failures instead of panicking. Indicating which functions may allocate is also a nice property. The best of both worlds would be keep these allocation results and optionally use the never type to gain back most of the performance by indicating the AllocErrors cannot be generated, but this only partially works in practice.

Partially gated by a new "alloc_error" feature, enabled by default. Disabling this feature and using nightly (along with the new "nightly" feature) will use the never type to compile out alloc errors and gain back some of the lost performance.

This will also help in safely catching OOMs in fuzzing.

## Tests

- All tests continue to pass
- Create a CI job to build and run in nightly with allocation errors turned off to test never type optimization