[runtime] Limit stack depth to prevent native stack overflows (#237)
## Summary

We are able to support JS stack frames up to the full size of the
managed JS stack with safe stack overflow checks. However the native
stack usually overflows before this point if a native stack frame is
added in recursive calls.

We enter the Rust runtime at many entrypoints from the VM making
tracking native stack depth hard. The best/easiest way to approximate
native stack depth is JS stack depth. So let's set a limit about 1/2 of
what's needed for some simple test cases. This is mean to be roughly
correct and should be refined as needed. Significantly higher limits can
be used in release builds.

When we exceed this max stack depth a regular stack overflow JS error is
thrown.

The fuzzer now runs without crashing on simple recursion that adds to
the native stack. This is the last known class of fuzzer false
positives.

## Tests

All tests pass.
- Added regression tests for some cases of native stack overflow from
fuzzer
- Ran fuzzer for a while, no longer see stack overflows due to recursion