[runtime] Fix spec compatibility bug leading to assertion failure in bitwise operations (#240)
## Summary

Fixes an assertion failure found by the fuzzer. We had a spec
compatibility bug where all binary bitwise or shift operations were
performing conversions twice by not using the output of the earlier call
to `to_numeric`.

This could lead to a crash as these second conversions are wrapped in
`must!`. If `to_number` were called on the output of `to_numeric` it
would be guaranteed to not throw but instead we perform the arbitrary
conversion. We initially observed this with a stack overflow error but
it can be reproduced by throwing any error.

## Tests

Added simple regression test for this case.