[runtime] must assertions propagate stack overflow errors instead of failing assertion (#241)
## Summary

Fixes a type of crash found by the fuzzer. Stack overflow errors are
treated as thrown JS values however these were mistakenly triggering
`must!` assertions to fail. This could occur in a `must!` that wraps a
call to a JS function that should not error (like resolve/reject) but
could still overflow the JS stack.

A stack overflow is special and we want to keep propagating it up. Let's
mark stack overflow errors with a new flag on `ErrorObject` so we can
distinguish them in the error path of a `must!` assertion.

Unfortunately we now need to split the `must!` macro into a special
version called within an `AllocResult` - `must_a!`.

## Tests

- All tests pass
- Run fuzzer some more, don't see this type of crash reappear