[runtime] Checked math in date calculations which may overflow (#242)
## Summary

The fuzzer caught assertion failures due to overflow during some date
calculations. Specifically callers to `make_day` could cause an overflow
if the year argument is high enough. We can fix this by using checked
operations in all functions that could potentially overflow.

This could only cause a crash due to the year argument - months and days
already have their ranges validated or controlled. I also audited the
rest of the date calculations and this was the only potential overflow I
found. All other cases operate on f64s or validate the range of their
inputs. The primary functions that rely on callers to validate inputs to
prevent overflow have been annotated with this fact.

## Tests

Added regression test for failing date calculations in the common cases
I could trigger it.