[runtime] Ensure initial realm stack frame is present and fix stack depth tracking edge cases (#243)
## Summary

Fixes a crash found by the fuzzer.

The first issue is that an initial stack frame is needed in order for
any native error objects to be created, since creating an error requires
accessing intrinsics through the current realm. We were failing to
guarantee this beforehand. Let's switch to a harder-to-misuse approach
where explicit push/pop of the initial stack frame are replaced by
closure passed to `Context::with_initial_realm_stack_frame`. We also
make sure that every entrypoint to execute code use this exactly once
(i.e. `run_script`, `run_modules`, and all task executors).

The second issue is that we had bugs in our explicit stack depth
tracking which was leading to unnecessary stack overflow errors. We now:
- Only increment the number of stack frames after verifying there is not
a stack overflow, otherwise stack depth is incremented but never
decremented
- Stop using `?` within the dispatch loop itself, instead use the new
`maybe_throw_a` utility when propagating an `AllocError`
- Debug assertion that stack + number of stack frames are empty when
pushing a new initial realm

## Tests

Added regression test which is a simplified version of the crash found
by the fuzzer.