[regexp] Parse legacy octal escape sequences in Annex B mode (#245)
## Summary

Implement the Annex B extension where legacy octal escape sequences can
appear in RegExps. We port the octal escape sequence parsing code over
from the main lexer.

The main complication is that in Annex B non-unicode aware mode, a
backreference can only be parsed if its capture group is a valid capture
index. Otherwise it is parsed as an escape sequence (which may be a
legacy octal escape sequence).

We do not know the total number of capture groups until after finishing
parsing, and don't currently have a good way of mutating parts of the
AST after parsing. Instead we keep track of each backreference while
parsing, and then afterwards determine if any of these are invalid and
reparse the entire RegExp if so. On this second parse we know the total
number of capture groups and what each backreference will parse to, so
we can detect invalid backreferences when they are encountered and
instead parse as an escape sequence.

## Tests

Now pass 5 more test262 Annex B tests:
- `built-ins/RegExp/RegExp-decimal-escape-class-range.js`
- `built-ins/RegExp/RegExp-decimal-escape-not-capturing.js`
- `built-ins/RegExp/RegExp-leading-escape.js`
- `built-ins/RegExp/RegExp-trailing-escape.js`
- `language/literals/regexp/legacy-octal-escape.js`