[bytecode] Fix clobber of for-in loop's lhs register by ForInNext result which signals the end of iteration (#247)
## Summary

Fixes a bug in bytecode generation where `ForInNext` was always
clobbering the variable of a for-in statement when the `ForInNext`
returns a result which signals the end of iteration. This is only
observable when the variable is a `var` that can be read after the loop.

The problem is that the `ForInNext` returns `undefined` when the loop
was done, but as an optimization we were storing this directly to the
variable's register destination (if stored directly in a register). We
should instead always be storing to a temporary first, checking if
`undefined`, then only writing back to the variable if the iteration
result was not `undefined`.

## Tests

- Added integration test verifying that for-in variable is not clobbered
after last iteration (or if there are no iterations)
- Update snapshot tests to see temporary register now used for
`ForInNext` result