[parser] For-in statements can have variable declaration intializer in Annex B mode (#248)
## Summary

Implement the Annex B extension where for-in statements can have an
initializer in their variable declaration. This initializer can only
occur in sloppy mode for statements of the form `for (var <id> = <expr>
in ...)`. Only allowed for `var` declarations with a single, simple
identifier pattern.

## Tests

All test262 `annexB/language/statements/for-in/*` tests now pass.